+++ lib/src/models/typing_indicator_mode.dart
@@ -1,0 +1,11 @@
+/// Used to toggle the display of avatars, names or both on [TypingIndicator].
+enum TypingIndicatorMode {
+  /// Show avatars of typing users.
+  avatar,
+
+  /// Show both avatars and names for typing users.
+  both,
+
+  /// Show names of typing users.
+  name,
+}