+++ lib/src/widgets/chat_list.dart
@@ -1,17 +1,20 @@
 import 'package:diffutil_dart/diffutil.dart';
-import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 
+import '../models/bubble_rtl_alignment.dart';
 import 'patched_sliver_animated_list.dart';
 import 'state/inherited_chat_theme.dart';
 import 'state/inherited_user.dart';
+import 'typing_indicator.dart';
 
 /// Animated list that handles automatic animations and pagination.
 class ChatList extends StatefulWidget {
   /// Creates a chat list widget.
   const ChatList({
     super.key,
+    this.bottomWidget,
+    required this.bubbleRtlAlignment,
     this.isLastPage,
     required this.itemBuilder,
     required this.items,
@@ -20,7 +23,16 @@ class ChatList extends StatefulWidget {
     this.onEndReachedThreshold,
     required this.scrollController,
     this.scrollPhysics,
+    this.typingIndicatorOptions,
+    required this.useTopSafeAreaInset,
   });
+
+  /// A custom widget at the bottom of the list.
+  final Widget? bottomWidget;
+
+  /// Used to set alignment of typing indicator.
+  /// See [BubbleRtlAlignment].
+  final BubbleRtlAlignment bubbleRtlAlignment;
 
   /// Used for pagination (infinite scroll) together with [onEndReached].
   /// When true, indicates that there are no more pages to load and
@@ -54,6 +66,13 @@ class ChatList extends StatefulWidget {
   /// Determines the physics of the scroll view.
   final ScrollPhysics? scrollPhysics;
 
+  /// Used to build typing indicator according to options.
+  /// See [TypingIndicatorOptions].
+  final TypingIndicatorOptions? typingIndicatorOptions;
+
+  /// Whether to use top safe area inset for the list.
+  final bool useTopSafeAreaInset;
+
   @override
   State<ChatList> createState() => _ChatListState();
 }
@@ -65,9 +84,9 @@ class _ChatListState extends State<ChatList>
     curve: Curves.easeOutQuad,
     parent: _controller,
   );
-
   late final AnimationController _controller = AnimationController(vsync: this);
 
+  bool _indicatorOnScrollStatus = false;
   bool _isNextPageLoading = false;
   final GlobalKey<PatchedSliverAnimatedListState> _listKey =
       GlobalKey<PatchedSliverAnimatedListState>();
@@ -97,6 +116,17 @@ class _ChatListState extends State<ChatList>
   Widget build(BuildContext context) =>
       NotificationListener<ScrollNotification>(
         onNotification: (notification) {
+          if (notification.metrics.pixels > 10.0 && !_indicatorOnScrollStatus) {
+            setState(() {
+              _indicatorOnScrollStatus = !_indicatorOnScrollStatus;
+            });
+          } else if (notification.metrics.pixels == 0.0 &&
+              _indicatorOnScrollStatus) {
+            setState(() {
+              _indicatorOnScrollStatus = !_indicatorOnScrollStatus;
+            });
+          }
+
           if (widget.onEndReached == null || widget.isLastPage == true) {
             return false;
           }
@@ -131,6 +161,21 @@ class _ChatListState extends State<ChatList>
           physics: widget.scrollPhysics,
           reverse: true,
           slivers: [
+            if (widget.bottomWidget != null)
+              SliverToBoxAdapter(child: widget.bottomWidget),
+            SliverPadding(
+              padding: const EdgeInsets.only(bottom: 4),
+              sliver: SliverToBoxAdapter(
+                child: widget.typingIndicatorOptions?.customTypingIndicator ??
+                    TypingIndicator(
+                      bubbleAlignment: widget.bubbleRtlAlignment,
+                      options: widget.typingIndicatorOptions!,
+                      showIndicator: (widget
+                              .typingIndicatorOptions!.typingUsers.isNotEmpty &&
+                          !_indicatorOnScrollStatus),
+                    ),
+              ),
+            ),
             SliverPadding(
               padding: const EdgeInsets.only(bottom: 4),
               sliver: PatchedSliverAnimatedList(
@@ -153,7 +198,10 @@ class _ChatListState extends State<ChatList>
             ),
             SliverPadding(
               padding: EdgeInsets.only(
-                top: 16 + (kIsWeb ? 0 : MediaQuery.of(context).padding.top),
+                top: 16 +
+                    (widget.useTopSafeAreaInset
+                        ? MediaQuery.of(context).padding.top
+                        : 0),
               ),
               sliver: SliverToBoxAdapter(
                 child: SizeTransition(
@@ -264,7 +312,7 @@ class _ChatListState extends State<ChatList>
         final message = item['message']! as types.Message;
 
         // Compare items to fire only on newly added messages.
-        if (oldMessage != message) {
+        if (oldMessage.id != message.id) {
           // Run only for sent message.
           if (message.author.id == InheritedUser.of(context).user.id) {
             // Delay to give some time for Flutter to calculate new